Create reusable select component
